Adding input pin reader callback
calling `analogRead()` directly in the class is problematic for people who are using an external I2C ADC board like the ADS1115. This change puts the input pin reading into a user definable callback. There is also a default method was just does an `analogRead` like the original library used.